feat: move fallback provider to use SealevelRpcClient instead
 - move higher level functions out of SealevelRpcClient and into
 SealevelProvider
