+++ rust/main/chains/hyperlane-sealevel/src/interchain_security_module.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 use num_traits::cast::FromPrimitive;
-use solana_sdk::{instruction::Instruction, pubkey::Pubkey};
+use solana_sdk::{instruction::Instruction, pubkey::Pubkey, signer::Signer};
 use tracing::warn;
 
 use hyperlane_core::{
@@ -12,20 +12,20 @@ use hyperlane_core::{
 use hyperlane_sealevel_interchain_security_module_interface::InterchainSecurityModuleInstruction;
 use serializable_account_meta::SimulationReturnData;
 
-use crate::{fallback::SealevelFallbackProvider, SealevelKeypair};
+use crate::{SealevelKeypair, SealevelProvider};
 
 /// A reference to an InterchainSecurityModule contract on some Sealevel chain
 #[derive(Debug)]
 pub struct SealevelInterchainSecurityModule {
     payer: Option<SealevelKeypair>,
     program_id: Pubkey,
-    provider: Arc<SealevelFallbackProvider>,
+    provider: Arc<SealevelProvider>,
 }
 
 impl SealevelInterchainSecurityModule {
     /// Create a new sealevel InterchainSecurityModule
     pub fn new(
-        provider: Arc<SealevelFallbackProvider>,
+        provider: Arc<SealevelProvider>,
         locator: ContractLocator,
         payer: Option<SealevelKeypair>,
     ) -> Self {
@@ -35,10 +35,6 @@ impl SealevelInterchainSecurityModule {
             program_id,
             provider,
         }
-    }
-
-    fn get_provider(&self) -> &SealevelFallbackProvider {
-        &self.provider
     }
 }
 
@@ -69,13 +65,14 @@ impl InterchainSecurityModule for SealevelInterchainSecurityModule {
             vec![],
         );
 
-        let payer = self
+        let pubkey = self
             .payer
             .as_ref()
+            .map(|p| p.pubkey())
             .ok_or_else(|| ChainCommunicationError::SignerUnavailable)?;
         let module = self
-            .get_provider()
-            .simulate_instruction::<SimulationReturnData<u32>>(payer.clone(), instruction)
+            .provider
+            .simulate_instruction::<SimulationReturnData<u32>>(&pubkey, instruction)
             .await?
             .ok_or_else(|| {
                 ChainCommunicationError::from_other_str("No return data was returned from the ISM")
