+++ rust/main/chains/hyperlane-sealevel/src/merkle_tree_hook.rs
@@ -24,6 +24,7 @@ impl MerkleTreeHook for SealevelMailbox {
 
         let outbox_account = self
             .get_provider()
+            .rpc_client()
             .get_account_with_finalized_commitment(self.outbox.0)
             .await?;
         let outbox = OutboxAccount::fetch(&mut outbox_account.data.as_ref())
