+++ rust/main/chains/hyperlane-sealevel/src/rpc/mod.rs
@@ -4,3 +4,5 @@ pub use client::SealevelRpcClient;
 pub mod client;
 /// SealevelRpcClientBuilder
 pub mod client_builder;
+/// rpc fallback client
+pub mod fallback;
