+++ rust/main/chains/hyperlane-sealevel/src/tx_submitter.rs
@@ -3,7 +3,8 @@ pub mod config;
 
 use std::sync::Arc;
 
-use config::TransactionSubmitterConfig;
+use async_trait::async_trait;
+
 use derive_new::new;
 use hyperlane_core::ChainResult;
 use solana_sdk::{
@@ -11,80 +12,88 @@ use solana_sdk::{
     signature::Signature, transaction::Transaction,
 };
 
-use crate::provider::fallback::SealevelFallbackProvider;
+use crate::SealevelProvider;
 
-/// The minimum tip to include in a transaction.
-/// From https://docs.jito.wtf/lowlatencytxnsend/#sendtransaction
-const JITO_MINIMUM_TIP_LAMPORTS: u64 = 1000;
-
-/// Transaction Submitter
-/// Configured to work with either Jito or Rpc
-#[derive(Clone, Debug, new)]
-pub struct TransactionSubmitter {
-    /// COnfig, for jito or rpc
-    pub config: TransactionSubmitterConfig,
-    /// provider
-    pub provider: Arc<SealevelFallbackProvider>,
-}
-
-impl TransactionSubmitter {
-    /// Get the RPC client
-    pub fn get_provider(&self) -> &SealevelFallbackProvider {
-        &self.provider
-    }
-
+/// A trait for submitting transactions to the chain.
+#[async_trait]
+pub trait TransactionSubmitter: Send + Sync {
     /// Get the instruction to set the compute unit price.
-    pub fn get_priority_fee_instruction(
+    fn get_priority_fee_instruction(
         &self,
         compute_unit_price_micro_lamports: u64,
         compute_units: u64,
         payer: &Pubkey,
-    ) -> Instruction {
-        match self.config {
-            TransactionSubmitterConfig::Jito { .. } => Self::jito_get_priority_fee_instruction(
-                compute_unit_price_micro_lamports,
-                compute_units,
-                payer,
-            ),
-            TransactionSubmitterConfig::Rpc { .. } => {
-                Self::rpc_get_priority_fee_instruction(compute_unit_price_micro_lamports)
-            }
-        }
-    }
+    ) -> Instruction;
 
     /// Send a transaction to the chain.
-    pub async fn send_transaction(
+    async fn send_transaction(
+        &self,
+        transaction: &Transaction,
+        skip_preflight: bool,
+    ) -> ChainResult<Signature>;
+
+    /// Get the RPC client
+    fn get_provider(&self) -> Option<&SealevelProvider> {
+        None
+    }
+}
+
+/// A transaction submitter that uses the vanilla RPC to submit transactions.
+#[derive(Debug, new)]
+pub struct RpcTransactionSubmitter {
+    provider: Arc<SealevelProvider>,
+}
+
+#[async_trait]
+impl TransactionSubmitter for RpcTransactionSubmitter {
+    fn get_priority_fee_instruction(
+        &self,
+        compute_unit_price_micro_lamports: u64,
+        _compute_units: u64,
+        _payer: &Pubkey,
+    ) -> Instruction {
+        ComputeBudgetInstruction::set_compute_unit_price(compute_unit_price_micro_lamports)
+    }
+
+    async fn send_transaction(
         &self,
         transaction: &Transaction,
         skip_preflight: bool,
     ) -> ChainResult<Signature> {
         self.provider
-            .call(move |provider| {
-                let transaction = transaction.clone();
-                let skip_preflight = skip_preflight;
-                let future = async move {
-                    provider
-                        .rpc_client()
-                        .send_transaction(&transaction, skip_preflight)
-                        .await
-                };
-                Box::pin(future)
-            })
+            .rpc_client()
+            .send_transaction(transaction, skip_preflight)
             .await
     }
 
-    fn rpc_get_priority_fee_instruction(compute_unit_price_micro_lamports: u64) -> Instruction {
-        ComputeBudgetInstruction::set_compute_unit_price(compute_unit_price_micro_lamports)
+    fn get_provider(&self) -> Option<&SealevelProvider> {
+        Some(&self.provider)
     }
+}
 
-    fn jito_get_priority_fee_instruction(
+/// A transaction submitter that uses the Jito API to submit transactions.
+#[derive(Debug, new)]
+pub struct JitoTransactionSubmitter {
+    provider: Arc<SealevelProvider>,
+}
+
+impl JitoTransactionSubmitter {
+    /// The minimum tip to include in a transaction.
+    /// From https://docs.jito.wtf/lowlatencytxnsend/#sendtransaction
+    const MINIMUM_TIP_LAMPORTS: u64 = 1000;
+}
+
+#[async_trait]
+impl TransactionSubmitter for JitoTransactionSubmitter {
+    fn get_priority_fee_instruction(
+        &self,
         compute_unit_price_micro_lamports: u64,
         compute_units: u64,
         payer: &Pubkey,
     ) -> Instruction {
         // Divide by 1_000_000 to convert from microlamports to lamports.
         let tip_lamports = (compute_units * compute_unit_price_micro_lamports) / 1_000_000;
-        let tip_lamports = tip_lamports.max(JITO_MINIMUM_TIP_LAMPORTS);
+        let tip_lamports = tip_lamports.max(Self::MINIMUM_TIP_LAMPORTS);
 
         // The tip is a standalone transfer to a Jito fee account.
         // See https://github.com/jito-labs/mev-protos/blob/master/json_rpc/http.md#sendbundle.
@@ -95,5 +104,16 @@ impl TransactionSubmitter {
             &solana_sdk::pubkey!("DfXygSm4jCyNCybVYYK6DwvWqjKee8pbDmJGcLWNDXjh"),
             tip_lamports,
         )
+    }
+
+    async fn send_transaction(
+        &self,
+        transaction: &Transaction,
+        skip_preflight: bool,
+    ) -> ChainResult<Signature> {
+        self.provider
+            .rpc_client()
+            .send_transaction(transaction, skip_preflight)
+            .await
     }
 }
