+++ rust/main/chains/hyperlane-sealevel/src/validator_announce.rs
@@ -11,29 +11,25 @@ use hyperlane_sealevel_validator_announce::{
 use solana_sdk::pubkey::Pubkey;
 use tracing::{info, instrument, warn};
 
-use crate::fallback::SealevelFallbackProvider;
+use crate::SealevelProvider;
 
 /// A reference to a ValidatorAnnounce contract on some Sealevel chain
 #[derive(Debug)]
 pub struct SealevelValidatorAnnounce {
-    provider: Arc<SealevelFallbackProvider>,
+    provider: Arc<SealevelProvider>,
     program_id: Pubkey,
     domain: HyperlaneDomain,
 }
 
 impl SealevelValidatorAnnounce {
     /// Create a new Sealevel ValidatorAnnounce
-    pub fn new(provider: Arc<SealevelFallbackProvider>, locator: &ContractLocator) -> Self {
+    pub fn new(provider: Arc<SealevelProvider>, locator: &ContractLocator) -> Self {
         let program_id = Pubkey::from(<[u8; 32]>::from(locator.address));
         Self {
             program_id,
             domain: locator.domain.clone(),
             provider,
         }
-    }
-
-    fn get_provider(&self) -> &SealevelFallbackProvider {
-        &self.provider
     }
 }
 
@@ -49,7 +45,7 @@ impl HyperlaneChain for SealevelValidatorAnnounce {
     }
 
     fn provider(&self) -> Box<dyn hyperlane_core::HyperlaneProvider> {
-        self.provider.provider()
+        Box::new(self.provider.clone())
     }
 }
 
@@ -77,7 +73,8 @@ impl ValidatorAnnounce for SealevelValidatorAnnounce {
         // Get all validator storage location accounts.
         // If an account doesn't exist, it will be returned as None.
         let accounts = self
-            .get_provider()
+            .provider
+            .rpc_client()
             .get_multiple_accounts_with_finalized_commitment(&account_pubkeys)
             .await?;
